+++ lib/posthog/registry.ex
@@ -1,12 +1,28 @@
 defmodule PostHog.Registry do
   @moduledoc false
   def config(supervisor_name) do
-    {:ok, config} =
-      supervisor_name
-      |> registry_name()
-      |> Registry.meta(:config)
+    registry = registry_name(supervisor_name)
 
-    config
+    if Process.whereis(registry) do
+      case Registry.meta(registry, :config) do
+        {:ok, config} -> config
+        :error -> disabled_config(supervisor_name)
+      end
+    else
+      disabled_config(supervisor_name)
+    end
+  rescue
+    ArgumentError -> disabled_config(supervisor_name)
+  end
+
+  def disabled_config(supervisor_name) do
+    %{
+      supervisor_name: supervisor_name,
+      enabled: false,
+      api_client: nil,
+      global_properties: %{},
+      test_mode: false
+    }
   end
 
   def registry_name(supervisor_name), do: Module.concat(supervisor_name, Registry)